Cleaned Main from unneccessery data and small fixes
