+++ pom.xml
@@ -95,7 +95,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>com.example.hrpulse/com.example.hrpulse.HR_Pulse</mainClass>
+                            <mainClass>com.example.hrpulse/com.example.hrpulse.DatabaseManager</mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
                             <jlinkImageName>app</jlinkImageName>
