+++ src/main/java/com/example/hrpulse/Services/Database/DatabaseManager.java
@@ -2,6 +2,8 @@ package com.example.hrpulse.Services.Database;
 
 import com.example.hrpulse.Services.Hibernate.HibernateUtil;
 import com.example.hrpulse.Services.Interfaces.DataModel;
+import com.example.hrpulse.Services.Objects.Department;
+import com.example.hrpulse.Services.Objects.Employee;
 import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
@@ -40,7 +42,138 @@ public class DatabaseManager {
     }
 
 
+    // Method to perform database operations after form submission for Employee
+    public static void performDatabaseOperations(Employee employee) {
+        DatabaseSessionManager sessionManager = new DatabaseSessionManager(DatabaseManager.getSessionFactory());
 
+        // Save the employee to the database
+        boolean saved = sessionManager.saveEmployee(employee);
+
+        if (saved) {
+            // Display confirmation
+            System.out.println("Employee saved successfully.");
+        } else {
+            // Display error
+            System.out.println("Error saving employee.");
+        }
+    }
+
+    // Method to retrieve a list of employees from the database
+    public static List<Employee> retrieveEmployees() {
+        try (Session session = sessionFactory.openSession()) {
+            session.beginTransaction();
+
+            // Create an HQL query to select all employees
+            Query<Employee> query = session.createQuery("from Employee", Employee.class);
+            List<Employee> employees = query.list();
+
+            // Commit the transaction
+            session.getTransaction().commit();
+            System.out.println("Retrieved " + employees.size() + " employees.");
+
+            return employees;
+        } catch (Exception e) {
+            e.printStackTrace();
+            System.err.println("Error retrieving employees: " + e.getMessage());
+            return null;
+        }
+    }
+
+    // Method to perform database operations after form submission for Department
+    public static void performDatabaseOperations(Department department) {
+        DatabaseSessionManager sessionManager = new DatabaseSessionManager(DatabaseManager.getSessionFactory());
+
+        // Save the department to the database
+        boolean saved = sessionManager.saveDepartment(department);
+
+        if (saved) {
+            // Display confirmation
+            System.out.println("Departments saved successfully.");
+        } else {
+            // Display error
+            System.out.println("Error saving Department.");
+        }
+    }
+
+    // Method to perform database operations after form submission for Department (update or remove)
+    public static void performDatabaseOperations(Department department, boolean update) {
+        DatabaseSessionManager sessionManager = new DatabaseSessionManager(DatabaseManager.getSessionFactory());
+
+        if (update) {
+            // Update the department in the database
+            boolean updated = sessionManager.updateDepartment(department);
+
+            if (updated) {
+                // Display confirmation
+                System.out.println("Department updated successfully.");
+            } else {
+                // Display error
+                System.out.println("Error updating Department.");
+            }
+        } else {
+            // Remove the department from the database
+            boolean removed = sessionManager.removeDepartment(department);
+
+            if (removed) {
+                // Display confirmation
+                System.out.println("Department removed successfully.");
+            } else {
+                // Display error
+                System.out.println("Error removing Department.");
+            }
+        }
+    }
+
+    // Method to retrieve a list of departments from the database
+    public static List<Department> retrieveDepartments() {
+        try (Session session = sessionFactory.openSession()) {
+            // Begin a transaction
+            session.beginTransaction();
+
+            // Create an HQL query to select all departments
+            Query<Department> query = session.createQuery("from Department", Department.class);
+
+            // Execute the query and get the list of departments
+            List<Department> departments = query.list();
+
+            // Commit the transaction
+            session.getTransaction().commit();
+
+            return departments;
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
+    }
+
+    // Method to perform database operations after form submission for Employee (update or remove)
+    public static void employeePDO(Employee selectedEmployee, boolean update) {
+        DatabaseSessionManager sessionManager = new DatabaseSessionManager(DatabaseManager.getSessionFactory());
+
+        if (update) {
+            // Update the employee in the database
+            boolean updated = sessionManager.updateEmployee(selectedEmployee);
+
+            if (updated) {
+                // Display confirmation
+                System.out.println("Employee updated successfully.");
+            } else {
+                // Display error
+                System.out.println("Error updating employee.");
+            }
+        } else {
+            // Remove the employee from the database
+            boolean removed = sessionManager.removeEmployee(selectedEmployee);
+
+            if (removed) {
+                // Display confirmation
+                System.out.println("Employee removed successfully.");
+            } else {
+                // Display error
+                System.out.println("Error removing Employee.");
+            }
+        }
+    }
 
 
     /**
